Split ops.py into torch / numpy backends with runtime dispatch
ops.py is now a thin loader+dispatcher: it loads the shared libs once, then
re-exports either _torch_ops (torch present) or _numpy_ops (torch-free). The
torch implementation is moved verbatim into _torch_ops.py, so the torch path is
byte-identical. _numpy_ops implements the video read path via the pybind
frontend returning numpy (DLPack), and raises a clear 'requires torch' for
audio/encode/wav/nvdec/file-like/compile.

pybind frontend gained the remaining read ops VideoDecoder needs:
get_frames_at_indices, get_frames_by_pts, get_frames_by_pts_in_range,
get_key_frame_indices, and add_video_stream now takes stream_index/num_threads.

Verified: torch path unchanged (test_decoders + test_ops 979 passed; re-exports
intact). This makes ops.py importable without torch; the remaining gate for a
torch-free 'import torchcodec' is guarding the package/decoders/_video_decoder
imports.
